Initial commit after splitting most of the module files.
