+++ geotools/geotools.py
@@ -4,6 +4,7 @@ from fuzzywuzzy import fuzz
 from unidecode import unidecode
 import tabletools
 
+from .utilities import *
 
 ####################################### Local Variables ############################################
 LOCAL_PATH = os.path.dirname(__file__)
