+++ systemtools/terminal.py
@@ -1,30 +1,4 @@
-import os
-import shlex
-import subprocess
-import psutil
 
-from . import timetools
-
-
-def memoryUsage(show = True, units = 'MB'):
-	""" Gets the current memory usage 
-		Returns
-		----------
-			if show is False
-			memory: int
-				The total number of bytes being used by the current process
-	"""
-
-	process = psutil.Process(os.getpid())
-	usage = process.memory_info().rss
-	if show:
-		if units == 'MB':
-			value = usage / 1024**2
-		else:
-			value = usage
-		print("Current memory usage: {0:.2f}{1}".format(value, units), flush = True)
-	else:
-		return usage
 
 
 class Terminal:
@@ -188,9 +162,3 @@ class Terminal:
 		except Exception as exception:
 			print("Could not write to the console file: ", str(exception))
 			print(filename)
-
-
-
-
-
-
