+++ tabletools/__init__.py
@@ -1,0 +1,2 @@
+from ._pandas_table import PandasTable as Table
+from .utilities import *
