+++ timetools/duration.py
@@ -1,175 +1,8 @@
+
 import datetime
 import re
 import time
 from numbers import Number
-
-# Import local files.
-try:
-	import numbertools
-except:
-	import pytools.numbertools as numbertools
-
-
-def elapsed(loop_number, loop_block, total_loops, timer):
-	""" Prints a line indicating the elapsed progress of the loop described
-		Parameters
-		----------
-			loop_number: int
-				The number of loops passed
-			loop_block: int
-				The number of loops to wait before updating the console
-			total_loops: int
-				The number of loops expected
-			timer: timetools.Timer
-				The timetools.Timer object used to track the current loop
-		Returns
-		----------
-			function : None
-		"""
-	_percent_complete = loop_number / total_loops
-	time_left = timer.togo(loop_number, total_loops, iso = True)
-	if loop_number % loop_block < 1:
-			print(
-				"{loop:<6}{complete:>6.1%} {left}".format(
-				loop = loop_number,
-				complete = _percent_complete,
-				left = time_left),
-				flush = True
-			)
-
-def now():
-	return Timestamp.now()
-
-class Timer:
-	def __init__(self, func = None, *args):
-		self.start_time = time.clock()
-		self.end_time = 0.0 
-		if func is not None:
-			self.timeFunction(func, 100, *args,)
-
-	def __str__(self):
-		return self.to_iso()
-
-	def __repr__(self):
-		return "Timer({0})".format(self.to_iso())
-
-	def duration(self):
-		self.end_time = time.clock()
-		return self.end_time - self.start_time
-
-	def isover(self, limit = 10.0):
-		""" Checks if more time has elapsed than the supplied limit
-			Parameters
-			----------
-				limit: int, float; default 10.0
-					The time limit, in seconds. 
-					If more than <limil> seconds have passed, returns True
-			Returns
-			----------
-				duration : bool
-		"""
-		result = self.duration() > limit
-		return result
-
-	def togo(self, done, total, iso = False):
-		""" Calculates the remaining time until a loop is finished
-			executing
-			Parameters
-			----------
-				done: int
-					The number of loops that have already passed
-				total: int
-					The total number of loops expected
-				iso: bool; default False
-					Whether to format the result as an ISO duration
-			Returns
-			----------
-				remaining : float, string
-		"""
-		if done == 0: done += 1
-		perloop = self.duration() / done
-		remaining = (total - done) * perloop
-		if iso:
-			remaining = Duration(remaining, 'Seconds').to_iso()
-		return remaining
-
-	def reset(self):
-		self.start_time = time.clock()
-
-	def split(self, label = 'the previous process'):
-		""" Prints the elapsed time, then resets the timer
-			Parameters
-			----------
-				label: string, int; default "the previous process"
-					Used to indicate which process was timed. If int, will
-					print the number of milliseconds required to run the 
-					previous loop instead
-			Returns
-			----------
-				function : None
-		"""
-		if isinstance(label, int):
-			string = self.timeit(label)
-		else:
-			string = "Finished {0} in {1}".format(label, self.to_iso())
-		print(string, flush = True)
-		self.reset()
-
-	def benchmark(self, loops = 1):
-		""" Returns a dictionary with information on the loop timing"""
-		duration = self.duration()
-		per_loop = duration / loops
-
-		result = {
-			'duration': duration,
-			'perLoop': per_loop,
-			'loops': loops
-		}
-		return result
-
-	def timeFunction(self, func, loops = 100, *args):
-		""" Benchmarks a function. Kwargs are passed on to the function.
-		"""
-		self.reset()
-		for i in range(loops):
-			func(*args)
-		self.timeit(loops)
-
-	def timeit(self, loops = 1, label = None):
-		""" Calculates the time for a loop to execute
-			Parameters
-			----------
-				loops: int
-					The total number of loops that were executed.
-				label: string
-					Added to the timeit message if provided.
-			Return
-			----------
-				duration : string
-		"""
-		benchmark = self.benchmark(loops)
-		duration = benchmark['duration']
-		per_loop = numbertools.humanReadable(benchmark['perLoop'])
-		
-		if label is None: message = ""
-		else: message = label + ': '
-		
-		message = message + "{0}s per loop ({1:.2f}s for {2:n} loop(s)) ".format(per_loop, duration, loops)
-		
-		print(message)
-		self.reset()
-		return message
-
-	def to_iso(self):
-		"""Returns an ISO duration of the elapsed time"""
-		seconds = self.duration()
-		return Duration(seconds, unit = 'Seconds').toiso()
-
-	def show(self, label = None):
-		if label is not None:   label += ': '
-		else:                   label = ''
-		print(label, "{0:.3f} seconds...".format(self.duration()), flush = True)
-
 
 class Duration(datetime.timedelta):
 	""" Inherits from datetime.timedelta. Designed to parse
@@ -497,181 +330,3 @@ class Duration(datetime.timedelta):
 			raise ValueError(message)
 
 		return value
-
-
-class Timestamp(datetime.datetime):
-	timestamp_regex = r"""(?:(?P<year>[\d]{4})-(?P<month>[\d]{2})-(?P<day>[\d]{2}))?
-						[\sA-Za-z]?
-						(?:(?P<hour>[\d]+)[:](?P<minute>[\d]+)[:](?P<second>[\d]+))?"""
-	timestamp_regex = re.compile(timestamp_regex, re.VERBOSE)
-
-	def __new__(cls, *args, **kwargs):
-		if len(args) == 1:
-			result = cls._parseInput(args[0])
-		elif len(args) == 3:
-			result = {'year': args[0], 'month': args[1], 'day': args[2]}
-			result.update(kwargs)
-		elif len(args) != 0:
-			result = args
-		else:
-			result = kwargs
-		if isinstance(result, dict):
-			return super().__new__(
-				cls,
-				result['year'], result['month'], result['day'],
-				result.get('hour', 0), result.get('minute', 0), result.get('second', 0))
-		else:
-			return super().__new__(cls, *result)
-
-	@staticmethod
-	def _cleandict(item):
-		item = {k: (int(v) if v else 0) for k, v in item.items()}
-		return item
-	
-	@classmethod
-	def _parseExcel(cls, value):
-		value += datetime.date(year = 1899, month = 12, day = 30).toordinal()
-			
-		xldate, xltime = divmod(value, 1)
-		date = datetime.date.fromordinal(int(xldate))
-		# ------------------Convert Time------------------
-		second = xltime * (3600 * 24)
-		second = int(second)
-		hour, second = divmod(second, 3600)
-		minute, second = divmod(second, 60)
-		
-		result = {
-			'year': date.year,
-			'month': date.month,
-			'day': date.day,
-			'hour': hour,
-			'minute': minute,
-			'second': second
-		}
-		return result
-
-	@classmethod
-	def _parseGenericObject(cls, value):
-		generic_date = cls._parseGenericDateObject(value)
-		generic_time = cls._parseGenericTimeObject(value)
-		if generic_date is None:
-			message = "Invalid Date: {}".format(value)
-			raise ValueError(message)
-		generic_date.update(generic_time)
-		return generic_date
-
-	@classmethod
-	def _parseGenericDateObject(cls, element):
-		try:
-			date_values = {
-				'year': element.year,
-				'month': element.month,
-				'day': element.day
-			}
-		except AttributeError:
-			date_values = None
-		return date_values
-
-	@classmethod
-	def _parseGenericTimeObject(cls, element):
-		try:
-			time_values = {
-				'hour': element.hour,
-				'minute': element.minute,
-				'second': element.second
-			}
-		except AttributeError:
-			time_values = dict()
-		return time_values
-
-	@classmethod
-	def _parseInput(cls, value):
-		if isinstance(value, str):
-			if any(not c.isdigit() for c in value):
-				result = cls._parseTimestamp(value)
-			else:
-				result = cls._parseNumericString(value)
-		elif isinstance(value, (tuple, list)):
-			result = cls._parseTuple(value)
-		else:
-			result = cls._parseGenericObject(value)
-
-		return result
-
-	@classmethod
-	def _parseDateTimeString(cls, string):
-		""" parses a string formatted as a generic YY/MM/DD string. """
-		if '-' in string and ':' in string:
-			result = cls._parseTimestamp(string)
-		else:
-			result = cls._parseNumericString(string)
-		return result
-
-	@classmethod
-	def _parseNumericString(cls, string):
-		""" Parses a date formatted as YY[YY]MMDD. """
-		string, day = string[:-2], string[-2:]
-		year, month = string[:-2], string[-2:]
-
-		year = int(year)
-		if year < 20: year += 2000
-		else: year += 1900
-
-		result = {
-			'year': int(year),
-			'month': int(month),
-			'day': int(day),
-			'hour': 0,
-			'minute': 0,
-			'second': 0
-		}
-		return result
-
-	@classmethod
-	def _parseTimestamp(cls, string):
-		""" Parses a date and/or time formated as YYYY-MM-DDThh:mm:ss"""
-		match = cls.timestamp_regex.search(string).groupdict()
-		match = cls._cleandict(match)
-		return match
-
-	@classmethod
-	def _parseTuple(cls, value):
-		keys = ('year', 'month', 'day', 'hour', 'minute', 'second')
-		result = dict(zip(keys, value))
-		return result
-
-	def getTime(self):
-		return self.hour, self.minute, self.second, self.microsecond
-
-	def toiso(self, compact = True):
-		result = self.isoformat()
-		if compact and not any(i != 0 for i in self.getTime()):
-			result = result.split('T')[0]
-		return result
-
-	def toNumeric(self):
-		pass
-
-	def toDict(self):
-		result = {
-			'year':   self.year,
-			'month':  self.month,
-			'day':    self.day,
-			'hour':   self.hour,
-			'minute': self.minute,
-			'second': self.second,
-			'microsecond': self.microsecond
-		}
-		return result
-
-	def toTuple(self):
-		result = (
-			self.year, self.month, self.day,
-			self.hour, self.minute, self.second, self.microsecond
-		)
-		return result
-
-if __name__ == "__main__":
-	timer = Timer()
-	time.sleep(2)
-	print(timer.to_iso())
